+++ R/datelife_query.R
@@ -20,6 +20,8 @@
 #'   given taxon or not. Default to `FALSE`. If `TRUE`, it must have same length as input.
 #'   If input is a newick string with some clades it will be converted to a `phylo`
 #'   object, and the order of `get_spp_from_taxon` will match `phy$tip.label`.
+#' @param taxonomic_source Used if `get_spp_from_taxon = TRUE`. A character vector with the desired taxonomic sources.
+#'  Options are "ott", "ncbi", "gbif" or "irmng". The function defaults to "ott".
 #' @return A `datelifeQuery` object, which is a list of three elements:
 #' \describe{
 #' 	 \item{$phy}{A `phylo` object or `NA`, if input is not a tree.}
@@ -33,15 +35,14 @@
 #' @export
 make_datelife_query <- function(input = c("Rhea americana", "Pterocnemia pennata", "Struthio camelus"),
                                 use_tnrs = TRUE,
-                                # approximate_match = TRUE,
-                                get_spp_from_taxon = FALSE) {
+                                get_spp_from_taxon = FALSE,
+                                taxonomic_source = "ott") {
   # enhance: add mapped (has tnrs been performed?) and matched (was it matched successfully?) element to phylo object
   # add one for each taxonomy queried: ott, catalogue of life (also contains fossils), enciclopedia of life (common names)
   if (suppressMessages(is_datelife_query(input))) {
     message("'input' is already a 'datelifeQuery' object.")
     return(input)
   }
-  message("... Making a DateLife query.")
   # input_process determines if input is newick and transforms it to phylo
   # if input is phylo or multiphylo it will also check if it is a good tree
   # if input is anything else, it will return NA:
@@ -50,6 +51,7 @@ make_datelife_query <- function(input = c("Rhea americana", "Pterocnemia pennata
   if (use_tnrs | any(get_spp_from_taxon)) {
     use_tnrs_global <- TRUE
   }
+  message("... Making a DateLife query:")
   if (inherits(phy_new, "phylo")) { # if input IS phylo
     cleaned_names <- phy_new$tip.label
     ott_ids <- NULL
@@ -73,8 +75,7 @@ make_datelife_query <- function(input = c("Rhea americana", "Pterocnemia pennata
   if (use_tnrs_global) {
     # process names even if it's a "higher" taxon name:
     cleaned_names_tnrs <- clean_tnrs(tnrs_match(input = cleaned_names),
-      remove_nonmatches = TRUE
-    )
+                                     remove_nonmatches = TRUE)
     # recover original names of invalid taxa and unmatched:
     ii <- !tolower(cleaned_names) %in% cleaned_names_tnrs$search_string
     cleaned_names <- c(cleaned_names_tnrs$unique_name, cleaned_names[ii])
@@ -97,40 +98,54 @@ make_datelife_query <- function(input = c("Rhea americana", "Pterocnemia pennata
       get_spp_from_taxon <- rep(get_spp_from_taxon, length(cleaned_names))
     }
     if (length(cleaned_names) != length(get_spp_from_taxon)) {
-      message("Specify all taxa in input to get species names from.")
+      message("Specify all taxa in input to get species names from -")
       message("'input' and 'get_spp_from_taxon' arguments must have same length.")
       return(NA)
     }
     # rotl::tol_subtree is very fast but returns subspecies too \o/
     # it has no argument to restrict it to species only
     # so we are using our own function that wraps up their services nicely
-    # example: df <- get_ott_children(ott_ids = 698424, ott_rank = "species")
-    df <- get_ott_children(ott_ids = cleaned_names_tnrs$ott_id, ott_rank = "species")
-    # head(rownames(df[[1]])[grepl("species", df[[1]]$rank)])
-    # the following does not work; it gives subspecies back
-    # fixing it from get_ott_children function and here too
-    cleaned_names <- lapply(df, function(x) rownames(x)[grepl("\\bspecies\\b", x$rank)])
-    # enhance: create vector original_taxon with original names: rep(cleaned_names[i], length(cleaned_names[i]))
-    original_taxa <- lapply(seq(nrow(cleaned_names_tnrs)), function(i) {
-      rep(cleaned_names_tnrs$unique_name[i], length(cleaned_names[[i]]))
-    })
-    ott_ids <- lapply(df, function(x) x$ott_id[grepl("\\bspecies\\b", x$rank)])
-    cleaned_names <- unlist(cleaned_names)
-    ott_ids <- unlist(ott_ids)
+    if ("ott" %in% taxonomic_source) {
+      species_list <- get_opentree_species(taxon_name = cleaned_names_tnrs$unique_name,
+                                           ott_id = cleaned_names_tnrs$ott_id,
+                                           synth_tree_only = TRUE)
+      cleaned_names <- species_list$species_names
+      ott_ids <- species_list$ott_ids
+    } else {
+      # example: df <- get_ott_children(ott_ids = 698424, ott_rank = "species")
+      df <- get_ott_children(ott_ids = cleaned_names_tnrs$ott_id, ott_rank = "species")
+      # head(rownames(df[[1]])[grepl("species", df[[1]]$rank)])
+      # the following does not work; it gives subspecies back
+      # fixing it from get_ott_children function and here too
+      cleaned_names <- lapply(df, function(x) rownames(x)[grepl("\\bspecies\\b", x$rank)])
+      # enhance: create vector original_taxon with original names: rep(cleaned_names[i], length(cleaned_names[i]))
+      original_taxa <- lapply(seq(nrow(cleaned_names_tnrs)), function(i) {
+        rep(cleaned_names_tnrs$unique_name[i], length(cleaned_names[[i]]))
+      })
+      ott_ids <- lapply(df, function(x) x$ott_id[grepl("\\bspecies\\b", x$rank)])
+      cleaned_names <- unlist(cleaned_names)
+      ott_ids <- unlist(ott_ids)
+   }
   }
+  # Make sure that we are using underscores and not spaces:
   cleaned_names <- gsub("_", " ", cleaned_names)
-  cleaned_names.print <- paste(cleaned_names, collapse = " | ")
-  message("Working with the following taxa: ", "\n", "\t", cleaned_names.print)
+  cleaned_names_print <- paste(utils::head(cleaned_names, 10), collapse = ", ")
   if (inherits(ott_ids, "numeric") | inherits(ott_ids, "integer")) {
     names(ott_ids) <- cleaned_names
   }
   # enhance: add original_taxa vector (from get_spp_from_taxon) to output here:
-  datelife_query.return <- list(
+  datelife_query_return <- list(
     cleaned_names = cleaned_names, ott_ids = ott_ids,
     phy = phy_new
   )
-  message("DateLife query done!")
-  return(structure(datelife_query.return, class = "datelifeQuery"))
+  #TODO: print working taxa to a file instead of screen:
+  message("Working with ",
+          length(cleaned_names),
+          " taxa: \n",
+          cleaned_names_print,
+          ifelse(length(cleaned_names) <=10, ".-", ", ..."))
+  message("DateLife query done!\n")
+  return(structure(datelife_query_return, class = "datelifeQuery"))
 }
 #' Process a phylo object or a character string to determine if it's correct newick
 #'
@@ -138,7 +153,7 @@ make_datelife_query <- function(input = c("Rhea americana", "Pterocnemia pennata
 #' @return A `phylo` object or `NA` if input is not a tree .
 #' @export
 input_process <- function(input) {
-  message("... Phylo-processing 'input'.")
+  message("... Phylo-processing 'input':")
   input_class <- "phylo"
   ott_ids <- NULL
   # TODO remove the multiPhylo if option from here?
@@ -162,19 +177,19 @@ input_process <- function(input) {
   if (any(grepl("\\(.*\\).*;", input))) { # our test for newick
     input <- input[grepl("\\(.*\\).*;", input)] # leave only the elements that are newick strings
     if (length(input) > 1) {
-      message("'input' has several newick strings. Only the first one will be used.")
+      message("'input' has several newick strings. Only the first one will be used...")
     }
     # phytools read.newick is not working
     # phy_out <- ape::collapse.singles(phytools::read.newick(text = gsub(" ", "_", input[1])))
     phy_out <- ape::collapse.singles(ape::read.tree(text = gsub(" ", "_", input[1])))
     phy_out$ott_ids <- ott_ids
     class(phy_out) <- input_class
-    message("'input' is a phylogeny and it is correctly formatted.")
+    message("'input' is a phylogeny and it is correctly formatted...")
     # ape::read.tree creates NaN edge lengths for tree without branch lengths
     # clean it up:
     if (!is.null(phy_out$edge.length)) {
       if (any(is.na(phy_out$edge.length))) {
-        warning("'input' has NA or NaN as branch lengths.")
+        warning("'input' has NA or NaN as branch lengths...")
         # phy_out$edge.length <- NULL
       }
     }